Updated to RunUO SVN 513.  Fixed the Carpentry issue where it would not use proper wood/logs.
Server files have been edited so new exe will be needed.

Please double check all the files to make sure you got them all before you post that it will not load or you are getting this error etc...

+++ Scripts/Mobiles/AI/Creature/BaseCreature.cs
@@ -250,10 +250,16 @@ namespace Server.Mobiles
 
         public virtual InhumanSpeech SpeechType { get { return null; } }
 
+        [CommandProperty(AccessLevel.GameMaster, AccessLevel.Administrator)]
         public bool IsStabled
         {
             get { return m_IsStabled; }
-            set { m_IsStabled = value; }
+            set
+            {
+                m_IsStabled = value;
+                if (m_IsStabled)
+                    StopDeleteTimer();
+            }
         }
 
         [CommandProperty(AccessLevel.GameMaster)]
@@ -365,12 +371,23 @@ namespace Server.Mobiles
 
         public void BeginDeleteTimer()
         {
-            if (!(this is BaseEscortable) && !Summoned && !Deleted)
+            if (!(this is BaseEscortable) && !Summoned && !Deleted && !IsStabled)
             {
+                StopDeleteTimer();
                 m_DeleteTimer = new DeleteTimer(this, TimeSpan.FromDays(3.0));
                 m_DeleteTimer.Start();
             }
         }
+
+        public void StopDeleteTimer()
+        {
+            if (m_DeleteTimer != null)
+            {
+                m_DeleteTimer.Stop();
+                m_DeleteTimer = null;
+            }
+        }
+
         #endregion
 
         public virtual double WeaponAbilityChance { get { return 0.4; } }
@@ -1350,24 +1367,24 @@ namespace Server.Mobiles
 
                 if (feathers != 0)
                 {
-                    corpse.DropItem(new Feather(feathers));
+                    corpse.AddCarvedItem(new Feather(feathers), from);
                     from.SendLocalizedMessage(500479); // You pluck the bird. The feathers are now on the corpse.
                 }
 
                 if (wool != 0)
                 {
-                    corpse.DropItem(new Wool(wool));
+                    corpse.AddCarvedItem(new Wool(wool), from);
                     from.SendLocalizedMessage(500483); // You shear it, and the wool is now on the corpse.
                 }
 
                 if (meat != 0)
                 {
                     if (MeatType == MeatType.Ribs)
-                        corpse.DropItem(new RawRibs(meat));
+                        corpse.AddCarvedItem(new RawRibs(meat), from);
                     else if (MeatType == MeatType.Bird)
-                        corpse.DropItem(new RawBird(meat));
+                        corpse.AddCarvedItem(new RawBird(meat), from);
                     else if (MeatType == MeatType.LambLeg)
-                        corpse.DropItem(new RawLambLeg(meat));
+                        corpse.AddCarvedItem(new RawLambLeg(meat), from);
 
                     from.SendLocalizedMessage(500467); // You carve some meat, which remains on the corpse.
                 }
@@ -1419,20 +1436,20 @@ namespace Server.Mobiles
 
                     switch (sc)
                     {
-                        case ScaleType.Red: corpse.DropItem(new RedScales(scales)); break;
-                        case ScaleType.Yellow: corpse.DropItem(new YellowScales(scales)); break;
-                        case ScaleType.Black: corpse.DropItem(new BlackScales(scales)); break;
-                        case ScaleType.Green: corpse.DropItem(new GreenScales(scales)); break;
-                        case ScaleType.White: corpse.DropItem(new WhiteScales(scales)); break;
-                        case ScaleType.Blue: corpse.DropItem(new BlueScales(scales)); break;
+                        case ScaleType.Red: corpse.AddCarvedItem(new RedScales(scales), from); break;
+                        case ScaleType.Yellow: corpse.AddCarvedItem(new YellowScales(scales), from); break;
+                        case ScaleType.Black: corpse.AddCarvedItem(new BlackScales(scales), from); break;
+                        case ScaleType.Green: corpse.AddCarvedItem(new GreenScales(scales), from); break;
+                        case ScaleType.White: corpse.AddCarvedItem(new WhiteScales(scales), from); break;
+                        case ScaleType.Blue: corpse.AddCarvedItem(new BlueScales(scales), from); break;
                         case ScaleType.All:
                             {
-                                corpse.DropItem(new RedScales(scales));
-                                corpse.DropItem(new YellowScales(scales));
-                                corpse.DropItem(new BlackScales(scales));
-                                corpse.DropItem(new GreenScales(scales));
-                                corpse.DropItem(new WhiteScales(scales));
-                                corpse.DropItem(new BlueScales(scales));
+                                corpse.AddCarvedItem(new RedScales(scales), from);
+                                corpse.AddCarvedItem(new YellowScales(scales), from);
+                                corpse.AddCarvedItem(new BlackScales(scales), from);
+                                corpse.AddCarvedItem(new GreenScales(scales), from);
+                                corpse.AddCarvedItem(new WhiteScales(scales), from);
+                                corpse.AddCarvedItem(new BlueScales(scales), from);
                                 break;
                             }
                     }
@@ -1633,7 +1650,10 @@ namespace Server.Mobiles
             #endregion
 
             // Version 18
-            writer.Write(DeleteTimeLeft);
+            if (IsStabled || (Controlled && ControlMaster != null))
+                writer.Write(TimeSpan.Zero);
+            else
+                writer.Write(DeleteTimeLeft);
         }
 
         private static double[] m_StandardActiveSpeeds = new double[]
@@ -2467,12 +2487,14 @@ namespace Server.Mobiles
             }
             set
             {
-                if (m_ControlMaster == value)
+                if (m_ControlMaster == value || this == value)
                     return;
 
                 RemoveFollowers();
                 m_ControlMaster = value;
                 AddFollowers();
+                if (m_ControlMaster != null)
+                    StopDeleteTimer();
 
                 Delta(MobileDelta.Noto);
             }
@@ -2487,7 +2509,7 @@ namespace Server.Mobiles
             }
             set
             {
-                if (m_SummonMaster == value)
+                if (m_SummonMaster == value || this == value)
                     return;
 
                 RemoveFollowers();
@@ -4275,11 +4297,6 @@ namespace Server.Mobiles
 
         public static List<DamageStore> GetLootingRights(List<DamageEntry> damageEntries, int hitsMax)
         {
-            return GetLootingRights(damageEntries, hitsMax, false);
-        }
-
-        public static List<DamageStore> GetLootingRights(List<DamageEntry> damageEntries, int hitsMax, bool partyAsIndividual)
-        {
             List<DamageStore> rights = new List<DamageStore>();
 
             for (int i = damageEntries.Count - 1; i >= 0; --i)
@@ -5287,6 +5304,7 @@ namespace Server.Mobiles
                 {
                     // *rummages through a corpse and takes an item*
                     PublicOverheadMessage(MessageType.Emote, 0x3B2, 1008086);
+                    //TODO: Instancing of Rummaged stuff.
                     return true;
                 }
             }
@@ -5316,7 +5334,10 @@ namespace Server.Mobiles
         {
             BardProvoked = true;
 
-            this.PublicOverheadMessage(MessageType.Emote, EmoteHue, false, "*looks furious*");
+            if (!Core.ML)
+            {
+                this.PublicOverheadMessage(MessageType.Emote, EmoteHue, false, "*looks furious*");
+            }
 
             if (bSuccess)
             {
